refactor: 비지니스 로직 domain 패키지로 이동
