+++ src/main/java/step5/domain/Car.java
@@ -1,4 +1,4 @@
-package step4.car;
+package step5.domain;
 
 public class Car {
