+++ src/main/java/step5/domain/CarFactory.java
@@ -1,4 +1,4 @@
-package step4.car;
+package step5.domain;
 
 import java.util.ArrayList;
 import java.util.Arrays;
