+++ src/main/java/step5/domain/RandomMovingStrategy.java
@@ -1,4 +1,4 @@
-package step4.car;
+package step5.domain;
 
 import java.util.Random;
