+++ src/main/java/step5/domain/Record.java
@@ -1,4 +1,4 @@
-package step4.car;
+package step5.domain;
 
 import java.util.List;
